Move ingestor out of core module
