+++ src/ingestor/errors.rs
